feat: Add form pre-fill for plantings form
The form is pre-filled using four keys: greenhouse/open field,
planting type, variety, crop. When one of these fours key is modified by
the user, we look for the latest planting which matches as much keys as
possible. Fields manually modified by the user won't be changed. We thus
have to track each fields (ChoiceChip, MyComboBox and MyTextField) for
modification.