fix: Fix bugs for plantings edit/update
We simplify the database schema, using the assigned date of the
planting's main planting task (sowing or transplanting) as planting_date.
We thus don't have to use triggers anymore to keep planting.planting_date
and task.assigned_date synchronized.  One of these triggers was
responsible for a strange date bug.