Show location value when in editing form
Instead of location ids, model indexes was passed to LocationDialog. So
it wasn't possible for LocationDialog to find the values.

We also changed the SQLite type of bed_width from INTEGER to FLOAT.
Since we now have a small userbase, we implemented a database migration
method and added our first migration script, 001.sql.