Add multiple database feature
It is now possible to work with two databases at the same time, and
switch between them by clicking on the button "1" and "2" on the left
panel. The user can create, open and export databases.

The location view will be reset when the user switches to another
database because the location tree must be rebuilt entirely.